Get core engine MUT passing
